+++ samples/net/rpl_node/boards/nrf52840_pca10056.conf
@@ -1,0 +1,4 @@
+CONFIG_IEEE802154_NRF5=y
+CONFIG_IEEE802154_NRF5_DRV_NAME="IEEE802154_nrf5"
+CONFIG_IEEE802154_NRF5_INIT_PRIO=80
+CONFIG_NET_APP_IEEE802154_DEV_NAME="IEEE802154_nrf5"
