+++ samples/net/rpl_node/boards/quark_se_c1000_devboard.conf
@@ -1,0 +1,11 @@
+CONFIG_ARC_INIT=n
+
+CONFIG_IEEE802154_CC2520=y
+
+CONFIG_NET_APP_IEEE802154_DEV_NAME="cc2520"
+
+CONFIG_IEEE802154_CC2520_RANDOM_MAC=n
+CONFIG_IEEE802154_CC2520_MAC4=0x00
+CONFIG_IEEE802154_CC2520_MAC5=0x00
+CONFIG_IEEE802154_CC2520_MAC6=0x00
+CONFIG_IEEE802154_CC2520_MAC7=0x01
