+++ samples/net/rpl_node/prj.conf
@@ -1,5 +1,3 @@
-CONFIG_ARC_INIT=n
-
 CONFIG_NETWORKING=y
 
 CONFIG_ENTROPY_GENERATOR=y
@@ -30,7 +28,6 @@ CONFIG_NET_LOG=y
 CONFIG_SYS_LOG_SHOW_COLOR=y
 CONFIG_SYS_LOG_NET_LEVEL=4
 CONFIG_INIT_STACKS=y
-CONFIG_PRINTK=y
 
 CONFIG_NET_BUF_LOG=n
 CONFIG_SYS_LOG_NET_BUF_LEVEL=1
@@ -44,28 +41,24 @@ CONFIG_NET_DEBUG_COAP=y
 
 CONFIG_NET_RPL=y
 CONFIG_NET_RPL_PROBING=y
-CONFIG_NET_RPL_STATS=y
-CONFIG_NET_RPL_MIN_HOP_RANK_INC=128
+CONFIG_NET_ROUTE=y
 
 CONFIG_NET_L2_IEEE802154=y
 CONFIG_NET_L2_IEEE802154_FRAGMENT=y
-CONFIG_IEEE802154_CC2520=y
-CONFIG_IEEE802154_CC2520_AUTO_ACK=n
 
 # IP core debug
-CONFIG_NET_DEBUG_CORE=y
-CONFIG_NET_DEBUG_IPV6_NBR_CACHE=y
-CONFIG_NET_DEBUG_IPV6=y
-CONFIG_NET_DEBUG_NET_BUF=n
+CONFIG_NET_DEBUG_CORE=n
+CONFIG_NET_DEBUG_IPV6_NBR_CACHE=n
+CONFIG_NET_DEBUG_IPV6=n
 CONFIG_NET_DEBUG_UTILS=n
-CONFIG_NET_DEBUG_IF=y
-CONFIG_NET_DEBUG_ICMPV6=y
+CONFIG_NET_DEBUG_IF=n
+CONFIG_NET_DEBUG_ICMPV6=n
 CONFIG_NET_DEBUG_CONN=n
 CONFIG_NET_DEBUG_UDP=n
 CONFIG_NET_DEBUG_CONTEXT=n
-CONFIG_NET_DEBUG_ROUTE=y
+CONFIG_NET_DEBUG_ROUTE=n
 CONFIG_NET_DEBUG_6LO=n
-CONFIG_NET_DEBUG_RPL=y
+CONFIG_NET_DEBUG_RPL=n
 
 CONFIG_NET_STATISTICS=n
 CONFIG_NET_STATISTICS_PERIODIC_OUTPUT=n
@@ -79,10 +72,3 @@ CONFIG_SYS_LOG_IEEE802154_DRIVER_LEVEL=0
 
 CONFIG_NET_APP_SETTINGS=y
 CONFIG_NET_APP_IEEE802154_CHANNEL=18
-CONFIG_NET_APP_IEEE802154_DEV_NAME="cc2520"
-
-CONFIG_IEEE802154_CC2520_RANDOM_MAC=n
-CONFIG_IEEE802154_CC2520_MAC4=0x00
-CONFIG_IEEE802154_CC2520_MAC5=0x00
-CONFIG_IEEE802154_CC2520_MAC6=0x00
-CONFIG_IEEE802154_CC2520_MAC7=0x03
