+++ samples/net/rpl_node/sample.yaml
@@ -1,14 +1,12 @@
 common:
   harness: net
-  depends_on: ieee802154
   tags: net rpl ieee802154
 sample:
   description: Can be used to test RPL network. Needs
     to be run in real device. This cannot be run in
     QEMU.
   name: RPL node demo application
+  platforms: all
 tests:
-  test_quark_se_c1000_devboard:
-    extra_args: CONF_FILE="prj_quark_se_c1000_devboard.conf"
-    platform_whitelist: quark_se_c1000_devboard
-    filter: ASSERT == 0
+  test:
+    platform_whitelist: quark_se_c1000_devboard nrf52840_pca10056
